requirements.txt
Django==2.1.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
pyhunter==0.3
clearbit==0.1.7
jsonfield==2.0.2
Faker==0.9.2
